VB error 438 invalid property or method
I get the following VB error in various parts.

```
Run-time error '438':

Object doesn't support this property or method
```

I didn't import the Dictionary module but instead copied the code from a module in another project - and in doing so lost the default property attribute.

This commit adds explicit Item calls in order to avoid this happening to other users.