New Windows-only Class WebAsyncClient
This is an alternative way of submitting asynchronous Web calls to
WenAsyncWrapper.

The WebClientAsync class truly wraps the WebClient class, providing the
same methods and properties and passing them through to WebClient. It is
inspired by WebAsyncWrapper and tries to make async calls as compatible
as possible with WebClient.

It also provides an ExecuteAsync method that kicks off a web call (using
another WebClientAsyncInstance class) and passes results or errors back
to the caller using Events.

Like WebAsyncWrapper it relies on WinHttpRequest and so is Windows only
(or at least until someone writes a WebHttpRequest compatible class for
Mac like Tim has done with the Dictionary class). Because it returns
results using Events, it can only be called from another Class module,
however it is the authors belief that anyone doing Async calls is likely
to be using classes anyway.

Having written and undertaken some simple tests, I am submitting this
code for early review. Once I have written the real code that will use
it and ironed out any bugs found, and rounded out the comments I will
let you know it is ready for merging.