Enforces UTF-8 decoding in WebResponse
WinHttpRequest uses the users current code page (CP_ACP) to decode the response body,
which on Windows is bound to fail for virtually all web services. This commit assumes the
response is in UTF-8, but to be correct it should be attempted to determine the encoding
from the Content-Type header followed by a fall-back to UTF-8. Also, the caller should have
a way of overriding the encoding. Also, maybe something along the lines of chardet could
be attempted.